+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.5.0
+- `getWineQualityDataFrame` function added
+- `loadIrisDataset`, `loadPimaIndiansDiabetesDataset` deprecated, `getIrisDataFrame`, `getPimaIndiansDiabetesDataFrame` added instead
+
 ## 1.4.2
 - `loadIrisDataset`, `loadPimaIndiansDiabetesDataset`: 
     - raw CSV data used instead of files